# [libc][math] Improve performance test framework