… libc/test/src/math/performance_testing/CMakeLists.txt
… libc/test/src/math/performance_testing/fmull_perf.cpp
… libc/test/src/math/performance_testing/hypot_perf.cpp
… libc/test/src/math/performance_testing/hypotf16_perf.cpp
… libc/test/src/math/performance_testing/hypotf_perf.cpp
… libc/test/src/math/performance_testing/log10f_perf.cpp
… libc/test/src/math/performance_testing/log1pf_perf.cpp
… libc/test/src/math/performance_testing/log2f_perf.cpp
… libc/test/src/math/performance_testing/logbf_perf.cpp
… libc/test/src/math/performance_testing/logf_perf.cpp
… libc/test/src/math/performance_testing/max_min_funcs_perf.cpp
… libc/test/src/math/performance_testing/misc_basic_ops_perf.cpp
… libc/test/src/math/performance_testing/nearbyintf_perf.cpp
… libc/test/src/math/performance_testing/nearest_integer_funcs_perf.cpp
… libc/test/src/math/performance_testing/rintf_perf.cpp
… libc/test/src/math/performance_testing/roundf_perf.cpp
… libc/test/src/math/performance_testing/sinf_perf.cpp
… libc/test/src/math/performance_testing/sqrtf128_perf.cpp
… libc/test/src/math/performance_testing/sqrtf_perf.cpp
… libc/test/src/math/performance_testing/truncf_perf.cpp