@meltq 2025-04-05
Old output
```
 Performance tests with inputs in denormal range:
-- My function --
     Total time      : 4545310 ns 
     Average runtime : 4443.12 ns/op 
     Ops per second  : 225067 op/s 
-- Other function --
     Total time      : 654400 ns 
     Average runtime : 639.687 ns/op 
     Ops per second  : 1563264 op/s 
-- Average runtime ratio --
     Mine / Other's  : 6.94577
```
New output
```
Function A - LIBC_NAMESPACE::hypotf16 Function B - LIBC_NAMESPACE::fputil::hypot<float16>
 Performance tests with inputs in denormal range:
-- Function A --
     Total time      : 2670736 ns 
     Average runtime : 2610.69 ns/op 
     Ops per second  : 383040 op/s 
-- Function B --
     Total time      : 499614 ns 
     Average runtime : 488.381 ns/op 
     Ops per second  : 2047580 op/s 
-- Average ops per second ratio --
     A / B  : 0.18707 
```