@overmighty 2025-04-15
I would prefer function names to be printed like this personally:

```
 Performance tests with inputs in denormal range:
-- Function A: LIBC_NAMESPACE::hypotf16 --
     Total time      : 2670736 ns
     Average runtime : 2610.69 ns/op
     Ops per second  : 383040 op/s
-- Function B: LIBC_NAMESPACE::fputil::hypot<float16> --
     Total time      : 499614 ns
     Average runtime : 488.381 ns/op
     Ops per second  : 2047580 op/s
-- Average ops per second ratio --
     A / B  : 0.18707
```